WIP: refactor frontend files + CSS
