+++ src/components/Login/index.tsx
@@ -1,9 +1,11 @@
 import React, { useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { legendary, loginPage, sidInfoPage } from '../helper'
+import { legendary, loginPage, sidInfoPage } from 'src/helpers'
 import Autorenew from '@material-ui/icons/Autorenew'
 import Info from '@material-ui/icons/Info'
-import LanguageSelector, { FlagPosition } from './UI/LanguageSelector'
+import LanguageSelector, {
+  FlagPosition,
+} from 'src/components/UI/LanguageSelector'
 const storage: Storage = window.localStorage
 interface Props {
   refresh: () => Promise<void>
