+++ src/components/Navbar/components/index.tsx
@@ -1,0 +1,2 @@
+export { default as SearchBar } from './SearchBar'
+export { default as UserSelector } from './UserSelector'
