+++ src/components/UI/GameCard.tsx
@@ -2,15 +2,15 @@
 import React, { useContext, useEffect, useState } from 'react'
 import { Link } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
-import ContextProvider from '../../state/ContextProvider'
-import { GameStatus } from '../../types'
-import { getProgress, sendKill, launch, updateGame } from '../../helper'
-import { handleInstall } from '../utls'
-import { ReactComponent as DownIcon } from '../../assets/down-icon.svg'
-import { ReactComponent as PlayIcon } from '../../assets/play-icon.svg'
-import { ReactComponent as StopIcon } from '../../assets/stop-icon.svg'
-import { ReactComponent as StopIconAlt } from '../../assets/stop-icon-alt.svg'
-import { ReactComponent as SettingsIcon } from '../../assets/settings-sharp.svg'
+import ContextProvider from 'src/state/ContextProvider'
+import { GameStatus } from 'src/types'
+import { getProgress, sendKill, launch, updateGame } from 'src/helpers'
+import { handleInstall } from '../utils'
+import { ReactComponent as DownIcon } from 'src/assets/down-icon.svg'
+import { ReactComponent as PlayIcon } from 'src/assets/play-icon.svg'
+import { ReactComponent as StopIcon } from 'src/assets/stop-icon.svg'
+import { ReactComponent as StopIconAlt } from 'src/assets/stop-icon-alt.svg'
+import { ReactComponent as SettingsIcon } from 'src/assets/settings-sharp.svg'
 import NewReleasesIcon from '@material-ui/icons/NewReleases'
 
 const { ipcRenderer, remote } = window.require('electron')
